This commit finishes with the second part of the equivalent values calculation. In this part of the procedure a query is performed looking for entries in the 'transactions' table that are missing a corresponding entry in the 'equivalent_values' table.
Whenere a transaction is detected to not have an equivalent value for a given fiat, a network request is made to the Coingecko's API in order
to obtain the BTS market value at the day of that transaction. An equivalent value is thus calculated using the 'bts_value' of the transaction.